Added ban/unban functionality
