+++ src/cmds/setup.js
@@ -1,3 +1,5 @@
+const Discord = require('discord.js');
+
 module.exports.run = (client, message, serverInfo, servers, sql, dev) => {
     if (message.member.hasPermission("ADMINISTRATOR") || message.author.id == dev) {
         message.channel.send(`Please provide me a prefix you would like to for use my commands, ${message.member}`).then(m => {
@@ -104,31 +106,4 @@ module.exports.run = (client, message, serverInfo, servers, sql, dev) => {
 
 
     }
-}
-
-
-
-function mysql_real_escape_string (str) {
-    return str.replace(/[\0\x08\x09\x1a\n\r"'\\\%]/g, function (char) {
-        switch (char) {
-            case "\0":
-                return "\\0";
-            case "\x08":
-                return "\\b";
-            case "\x09":
-                return "\\t";
-            case "\x1a":
-                return "\\z";
-            case "\n":
-                return "\\n";
-            case "\r":
-                return "\\r";
-            case "\"":
-            case "'":
-            case "\\":
-            case "%":
-                return char+char; // prepends a backslash to backslash, percent,
-                                  // and double/single quotes
-        }
-    });
 }
